Cleanup and formatting
Removed an ugly hack that was supposed to prevent quest building assignments from being overwritten by each other, but it didn't handle cases where the quest was taken, failed, or in progress, which are states I'd like to be able to make use of. This hack can all be avoided by just managing quest states properly.