Implemented Devouring bag curse
Items cant be 't'aken out of, 'd'ropped from, or 'v'Thrown from a Devouring bag.

Despite the name, this curse does not destroy or eat any items, but merely turns your bag into a Deposit Only drop box.

Note that 'E'ating, 'q'uaffing, 'r'eading, 'F'ueling, 'k'destroying, and '{'Inscribing items inside such a cursed bag does still work, which I think is reasonable for the lore of the bag and reduces the number of functions that need to be patched.

There may be other functions that interact with bags which I haven't thought of though.